refactor: extracted interface OpenInsider and renamed implementation to OpenInsiderScraper
